Constans moved to platform
